Code cleaning

+++ Source/ServerTime.cs
@@ -10,7 +10,7 @@ public class ServerTime : MonoBehaviour
 {
     private void OnGUI()
     {
-        GUILayout.BeginArea(new Rect((float) ((Screen.width / 2) - 100), 0f, 200f, 30f));
+        GUILayout.BeginArea(new Rect((Screen.width / 2) - 100, 0f, 200f, 30f));
         GUILayout.Label(string.Format("Time Offset: {0}", PhotonNetwork.networkingPeer.ServerTimeInMilliSeconds - Environment.TickCount), new GUILayoutOption[0]);
         if (GUILayout.Button("fetch", new GUILayoutOption[0]))
         {
